Cleaned up color picker/button + fixed some bugs with settings context updating and styling
